fixed EnumHelper Bug cleanedup codes and files
