+++ Plugins/BulkExporterPlugin/Enums/EnumHelper.cs
@@ -1,10 +1,10 @@
-﻿using FrostySdk.Attributes;
-using System;
+﻿using System;
 using System.Collections.Generic;
 using System.Globalization;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using System.ComponentModel;
 
 namespace BulkExporterPlugin.Enums
 {
@@ -17,9 +17,7 @@ namespace BulkExporterPlugin.Enums
                 return (attributes.First() as DescriptionAttribute).Description;
 
             // If no description is found, the least we can do is replace underscores with spaces
-            // You can add your own custom default formatting logic here
-            TextInfo ti = CultureInfo.CurrentCulture.TextInfo;
-            return ti.ToTitleCase(ti.ToLower(value.ToString().Replace("_", " ")));
+            return value.ToString().Replace("_", " ");
         }
 
         public static IEnumerable<ValueDescription> GetAllValuesAndDescriptions(Type t)
@@ -27,7 +25,7 @@ namespace BulkExporterPlugin.Enums
             if (!t.IsEnum)
                 throw new ArgumentException($"{nameof(t)} must be an enum type");
 
-            return Enum.GetValues(t).Cast<Enum>().Select((e) => new ValueDescription() { Value = e, Description = e.Description() }).ToList();
+            return Enum.GetValues(t).Cast<Enum>().Select((e) => new ValueDescription() { Value = e, Description = e.Description() }).ToList();;
         }
 
         public class ValueDescription
