--- Plugins/BulkExporterPlugin/Enums/MeshFormat.cs
@@ -1,12 +1,0 @@
-﻿using System.ComponentModel;
-
-namespace BulkExporterPlugin.Enums
-{
-    public enum MeshFormat
-    {
-        [Description(".FBX")]
-        FBX,
-        [Description(".OBJ")]
-        OBJ
-    }
-}
