--- Plugins/BulkExporterPlugin/Enums/MeshScale.cs
@@ -1,11 +1,0 @@
-﻿
-namespace BulkExporterPlugin.Enums
-{
-    public enum MeshScale
-    {
-        Millimeters,
-        Centimeters,
-        Meters,
-        Kilometers
-    }
-}
