--- Plugins/BulkExporterPlugin/Enums/MeshVersion.cs
@@ -1,13 +1,0 @@
-﻿
-namespace BulkExporterPlugin.Enums
-{
-    public enum MeshVersion
-    {
-        FBX_2012,
-        FBX_2013,
-        FBX_2014,
-        FBX_2015,
-        FBX_2016,
-        FBX_2017
-    }
-}
