--- Plugins/BulkExporterPlugin/Enums/TextureFormat.cs
@@ -1,10 +1,0 @@
-﻿namespace BulkExporterPlugin.Enums
-{
-    public enum TextureFormat
-    {
-        PNG,
-        HDR,
-        TGA,
-        DDS
-    }
-}
