+++ Plugins/BulkExporterPlugin/Themes/Generic.xaml
@@ -1,7 +1,6 @@
 ﻿<ResourceDictionary 
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:local="clr-namespace:BulkExporterPlugin"
     xmlns:core="clr-namespace:Frosty.Core.Controls;assembly=FrostyCore"
     xmlns:ctrl="clr-namespace:Frosty.Controls;assembly=FrostyControls"
     xmlns:editors="clr-namespace:BulkExporterPlugin.Editors"
@@ -14,10 +13,10 @@
     </ResourceDictionary.MergedDictionaries>
 
     <!-- BulkExporter -->
-    <Style TargetType="{x:Type local:BulkExporter}">
+    <Style TargetType="{x:Type editors:BulkExporter}">
         <Setter Property="Template">
             <Setter.Value>
-                <ControlTemplate TargetType="{x:Type local:BulkExporter}">
+                <ControlTemplate TargetType="{x:Type editors:BulkExporter}">
                     <ControlTemplate.Resources>
                         
                     </ControlTemplate.Resources>
