+++ Plugins/BulkExporterPlugin/Windows/BulkExportWindow.xaml.cs
@@ -6,18 +6,10 @@ using System;
 using System.Collections.Generic;
 using System.Collections.ObjectModel;
 using System.ComponentModel;
-using System.Linq;
 using System.Runtime.CompilerServices;
 using System.Text;
 using System.Threading.Tasks;
 using System.Windows;
-using System.Windows.Controls;
-using System.Windows.Data;
-using System.Windows.Documents;
-using System.Windows.Input;
-using System.Windows.Media;
-using System.Windows.Media.Imaging;
-using System.Windows.Shapes;
 
 namespace BulkExporterPlugin.Windows
 {
@@ -27,6 +19,7 @@ namespace BulkExporterPlugin.Windows
     public partial class BulkExportWindow : FrostyMessageBox, INotifyPropertyChanged
     {
         private BulkExportSetting _setting;
+        public BulkExportSetting ExportSetting { get => _setting; }
 
         #region --Flatten--
         public bool Flatten
